requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
reportlab==4.0.7
django==4.2.7
django-ninja==1.0.1
python-dotenv==1.0.0
alembic==1.12.1
jinja2==3.1.2
weasyprint==60.1 